+++ source/example/F0.Analyzers.Example.CSharp7/CodeAnalysis/CodeRefactorings/ObjectInitializerExample.cs
@@ -6,7 +6,7 @@ namespace F0.Analyzers.Example.CSharp7.CodeAnalysis.CodeRefactorings
 	[System.Diagnostics.CodeAnalysis.SuppressMessage("Style", "IDE0022:Use expression body for methods", Justification = "Examples")]
 	internal sealed class ObjectInitializerExample
 	{
-		public Model GetModel()
+		public Model DefaultLiteralIsNotAvailableInCSharp7()
 		{
 			return new Model();
 		}