+++ source/example/F0.Analyzers.Example/Shared/BaseClass.cs
@@ -1,0 +1,13 @@
+﻿namespace F0.Analyzers.Example.Shared
+{
+	internal abstract class BaseClass : IInterface
+	{
+		public int Field;
+
+		int IInterface.Length { get; set; }
+		public int Current { get; set; }
+		public int Value { get; set; }
+
+		public abstract int Property { get; set; }
+	}
+}