+++ CHANGELOG.md
@@ -2,10 +2,11 @@
 CHANGELOG
 
 ## vNext
+- Changed `ObjectInitializer` code refactoring, additionally assigning inherited members from `base` classes.
 
 ## v0.5.0 (2020-09-29)
 ### Analyzers
-- Changed `ObjectInitializer` code refactoring, assigning `internal` fields and properties within files in the same assembly or _friend_ assemblies.
+- Changed `ObjectInitializer` code refactoring, additionally assigning `internal` fields and properties within files in the same assembly or _friend_ assemblies.
 
 ## v0.4.1 (2020-07-11)
 ### Analyzers